Fix building pkg/util/procfs on non-linux platforms
