+++ pkg/util/procfs/procfs_linux_test.go
@@ -1,3 +1,5 @@
+// +build linux
+
 /*
 Copyright 2015 The Kubernetes Authors.
